+++ tp4/ParticleConsts.cs
@@ -2,7 +2,7 @@ namespace tp4;
 
 public struct ParticleConsts
 {
-    public float Mass { get; set; }
+    public double Mass { get; set; }
 
     public override string ToString()
     {
